Adding a logCommand method
Adding a method that will log a command and all output to standard out.  This feature is not yet ready to be released in the public API, as it has not yet been thoroughly tested.  Even if it were, though, handling scope for all but the simplest use cases makes this method not very elegant to use.

To log a simple command:
logger.logCommand('fprintf(''Logging output'')');

To log a simple command with output variables assigned:
[a, b, c] = logger.logCommand('deal(1)');

Scope is not shared when you call logCommand.  So if you need to share scope, you must pass the scope into the logCommand method.  You can do so as a string:

x = logger.logCommand( ...
    'fmincon(fun,x0,A,b)', ...
    'fun = @(x)100*(x(2)-x(1)^2)^2, x0 = [-1,2], A = [1,2], b = 1' ...
);

Or as a list of variables:

fun = @(x)100*(x(2)-x(1)^2)^2 + (1-x(1))^2;
x0 = [-1,2];
A = [1,2];
b = 1;

x = logger.logCommand('fmincon(fun,x0,A,b)', fun, x0, A, b);